organized tests and added and updated tests related to the floor system
